// src/app/core/models/vehicule-societe.ts
/**
 * Modèle représentant un véhicule de société
 * 
 */
export interface VehiculeSociete {
    id?: number;
    immatriculation?: string;
    marque?: string;
    modele?: string;
    places?: number;
    photo?: string;
    disponible?: boolean;
    statut?: number;
    categorie?: string;
}